+++ src/fixIcelandicLocale.test.ts
@@ -418,9 +418,16 @@ describe('_PatchedListFormat', () => {
       (['long', 'short', 'narrow'] as const).forEach((style) => {
         [['a', 'b', 'c'], ['a', 'b'], ['a']].forEach((input) => {
           const opts = { style, type };
-          expect(new _PatchedListFormat('is', opts).format(input)).toEqual(
-            new _PatchedListFormat.$original('is', opts).format(input)
+          const narrowUnit = style === 'narrow' && type === 'unit';
+          const comma = narrowUnit ? '' : ',';
+          const joint = narrowUnit ? '' : type === 'disjunction' ? ' eða' : ' og';
+          expect(new _PatchedListFormat('is', opts).format(['a', 'b', 'c'])).toEqual(
+            `a${comma} b${joint} c`
           );
+          expect(new _PatchedListFormat('is', opts).format(['a', 'b'])).toEqual(
+            `a${joint} b`
+          );
+          expect(new _PatchedListFormat('is', opts).format(['a'])).toEqual(`a`);
         });
       });
     });